Improved dine-out request functionality.
Prevented NullPointerException when attempting to make a request without
any requested items, added functionality to incorporate the user's
additional notes in the request, and added functionality to return the
user to the main menu screen upon a successful request.

Also added functionality to return the user to the main menu screen upon
a delayed success while making a reservation.